Node.js での認証状態の管理を各種ファイルに入力しなおし
